subscriber/requirements.txt
paho-mqtt==1.4.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
